+++ packages/dht/src/connection/connectivityRequestHandler.ts
@@ -5,7 +5,7 @@ import {
     Message, MessageType
 } from '../proto/packages/dht/protos/DhtRpc'
 import { NatType } from './ConnectionManager'
-import { ConnectivityChecker, connectAsync } from './ConnectivityChecker'
+import { CONNECTIVITY_CHECKER_SERVICE_ID, connectAsync } from './connectivityChecker'
 import { IConnection } from './IConnection'
 import { ServerWebsocket } from './websocket/ServerWebsocket'
 import { connectivityMethodToWebsocketUrl } from './websocket/WebsocketConnector'
@@ -66,7 +66,7 @@ const handleIncomingConnectivityRequest = async (connection: ServerWebsocket, co
         }
     }
     const msg: Message = {
-        serviceId: ConnectivityChecker.CONNECTIVITY_CHECKER_SERVICE_ID,
+        serviceId: CONNECTIVITY_CHECKER_SERVICE_ID,
         messageType: MessageType.CONNECTIVITY_RESPONSE,
         messageId: v4(),
         body: {
